+++ CHANGELOG.md
@@ -2,6 +2,7 @@
  - Custom enterprise certificates creation fix (now sha256) for modern hostapd-wpe versions
  - Fixed Docker distro shown (now Kali based)
  - Fixed PMKID hashcat hash parsing
+ - Improvements on graphics system and resolution detection
 
 ### 11.10
  - Fixed problem for latest BeEF